Configuration's default values equals API defaults.
Meaning we set units and mode to metric and driving.
At the same time we don't include these parameters
if for instance the mode is driving, as that is the
default we get back if we don't include mode in the
list og GET params.

In effect we can ask our configuration object what
type of metric is being used, but don't clutter the
URL (which has a limit of 2048 chars) with params
which are not needed.