UrlBuilder knows how to use polyline encoder.
Places are encoded in to a polyline when we have consecutive places
which has lat/lng. If I remember correctly the order of given
destination/origins matters, so we cannot do all places with lat/lng in
one long line, and then address places thereafter. They have to be
broken off and be in correct order.